Node/requirements.txt
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
func-timeout==4.3.5
itsdangerous==1.1.0
Jinja2==2.11.2
kademlia==2.2.1
MarkupSafe==1.1.1
peewee==3.13.3
pycparser==2.20
pycryptodome==3.9.8
pyOpenSSL==19.1.0
rpcudp==4.0.1
six==1.15.0
u-msgpack-python==2.6.0
Werkzeug==1.0.1
